Commit : Ajout des méthodes RecupérerForfait et RécupérerHorsForfait

+++ AP_1_GSB/Data/Models/FraisHorsForfait.cs
@@ -10,7 +10,7 @@ namespace AP_1_GSB.Data.Models
     {
         public int IdFraisForfait { get; set; }
         public string Description { get; set; }
-        public int Montant { get; set; }
+        public float Montant { get; set; }
         public DateTime Date { get; set; }
         public EtatFraisHorsForfait Etat { get; set; }
         public Justificatif Justificatif { get; set; }
